# docs/plans/api-specification.yaml
openapi: 3.1.0
info:
  title: ekaiX AIXcelerator API
  version: 1.0.0
  description: |
    Backend API for ekaiX — a Snowflake Native App that enables business users to create
    semantic models and Cortex Agents through AI conversation.

    **Authentication:** All requests are authenticated via the `Sf-Context-Current-User` header
    injected by SPCS (Snowpark Container Services). No separate auth system exists.

    **Workspace Isolation:** All data is scoped to the authenticated user's workspace via
    PostgreSQL Row-Level Security. Sharing is explicit via the shares endpoints.

    **Streaming:** Agent responses are delivered via Server-Sent Events (SSE) on the
    `/agent/stream/{session_id}` endpoint.
  contact:
    name: ekai Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://{spcs_endpoint}
    description: SPCS production
    variables:
      spcs_endpoint:
        default: ekaix.snowflakecomputing.app
...
security:
  - spcsAuth: []
components:
  securitySchemes:
    spcsAuth:
      type: apiKey
      in: header
      name: Sf-Context-Current-User
      description: Snowflake username injected by SPCS. Cannot be spoofed in production.

  # ==========================================================================
  # SCHEMAS
  # ==========================================================================
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Error code (e.g., NOT_FOUND, VALIDATION_ERROR, FORBIDDEN)
          examples: ["NOT_FOUND"]
        message:
          type: string
          description: Human-readable error description
          examples: ["Data product not found"]
        details:
          type: object
          description: Additional error context (validation errors, field-level issues)
          additionalProperties: true
...
    UploadedDocument:
      type: object
      required: [id, filename, extraction_status, created_at]
      properties:
        id:
          type: string
          format: uuid
        data_product_id:
          type: string
          format: uuid
        filename:
          type: string
        file_size_bytes:
          type: integer
        content_type:
          type: string
        extraction_status:
          type: string
          enum: [pending, processing, completed, failed]
        extraction_error:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        extracted_at:
          type: string
          format: date-time
          nullable: true
    DocumentContent:
      type: object
      required: [id, filename, content]
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        content:
          type: string
          description: Extracted text content from Cortex Document AI
        extraction_status:
          type: string

    # --- Health Check ---
...
paths:
...
  /documents/upload:
    post:
      tags: [Documents]
      operationId: uploadDocument
      summary: Upload a PDF or DOCX document
      description: |
        Uploads a document to MinIO and optionally triggers Cortex Document AI extraction.
        Max file size: 50MB. Supported types: application/pdf, application/vnd.openxmlformats-officedocument.wordprocessingml.document.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [data_product_id, file]
              properties:
                data_product_id:
                  type: string
                  format: uuid
                file:
                  type: string
                  format: binary
                auto_extract:
                  type: boolean
                  default: true
                  description: Automatically trigger text extraction after upload
      responses:
        "201":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadedDocument"
        "400":
          description: Invalid file type or size exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: File too large (max 50MB)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /documents/{id}/content:
    get:
      tags: [Documents]
      operationId: getDocumentContent
      summary: Get extracted text content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Extracted document content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentContent"
        "404":
          description: Document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Extraction not yet completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /documents/{id}:
    delete:
      tags: [Documents]
      operationId: deleteDocument
      summary: Delete an uploaded document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Document deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"